Cleanup: split core.views into more manageable modules
Most notably decorators and mixins have been put apart. For the latter the
import has been adjusted to `core.views.mixins`; forcing developers to use it
serves as a hint of a word of warning.
