+++ pootle/apps/pootle_app/views/admin/languages.py
@@ -8,7 +8,8 @@
 
 from django.views.generic import TemplateView
 
-from pootle.core.views import APIView, SuperuserRequiredMixin
+from pootle.core.views import APIView
+from pootle.core.views.mixins import SuperuserRequiredMixin
 from pootle_app.forms import LanguageForm
 from pootle_language.models import Language
